ログデータ一覧取得のテスト関数を追加。また、SimpleStatusのユーザIDをint型からlong型に変更。
git-svn-id: http://svn.sourceforge.jp/svnroot/nt-manager/NishioTweetManager/trunk@132 d8c9ecd3-d47d-4367-8645-de82c00e513f

+++ src/twitter/manage/StatusXMLConverter.java
@@ -414,7 +414,7 @@ public class StatusXMLConverter {
 					status.setInReplyToStatusId(Long.parseLong(replyStatusID));
 				}
 				if (replyUserID != null) {
-					status.setInReplyToUserId(Integer.parseInt(replyUserID));
+					status.setInReplyToUserId(Long.parseLong(replyUserID));
 				}
 				// TODO: Place
 				// TODO: RetweetStatus ここは実装可能
